Move documentation out of /thingpedia/developers and into /doc
Documentation includes Almond documentation (user & developers),
and therefore it is useful to have even if Thingpedia is disabled.
