+++ public/javascripts/docsearch.js
@@ -14,7 +14,7 @@ $(function() {
             return;
         }
             
-        $.ajax('/thingpedia/developers/search', { data: { q: q } }).then(function(results) {
+        $.ajax('/doc/search', { data: { q: q } }).then(function(results) {
             container.empty();
             results.data.forEach(function(result) {
                 var item = $('<a>').addClass('list-group-item').attr('href', result.url).html('<span>' + result.highlight + '</span>');
