+++ routes/thingpedia_developer_console.js
@@ -10,9 +10,6 @@
 "use strict";
 
 const express = require('express');
-const path = require('path');
-const lunr = require('lunr');
-require("lunr-languages/lunr.stemmer.support")(lunr);
 
 const db = require('../util/db');
 const organization = require('../model/organization');
@@ -260,63 +257,11 @@ router.get('/status', (req, res) => {
     res.redirect('/me/status');
 });
 
-for (let doc of require('../doc/doc-list.json')) {
-    router.get('/' + doc + '.md', (req, res, next) => {
-        res.render('doc_' + doc, {
-            page_title: req._("Thingpedia - Documentation"),
-            currentPage: doc
-        });
-    });
-}
-
-const searchIndex = require('../doc/fts.json');
-searchIndex.index = lunr.Index.load(searchIndex.index);
-function highlightSearch(url, metadata) {
-    const terms = [];
-    let minIndex = Infinity;
-    let maxIndex = -Infinity;
-    for (let term in metadata) {
-        if (metadata[term].content) {
-            terms.push(term);
-            for (let pos of metadata[term].content.position) {
-                minIndex = Math.min(pos[0], minIndex);
-                maxIndex = Math.max(pos[1]+pos[0], maxIndex);
-            }
-        }
-    }
-    
-    const content = searchIndex.documents[url].content;
-    if (!terms.length)
-        return content;
-    
-    const trimLeft = minIndex > 10;
-    
-    const trimmedText = (trimLeft ? '...' : '') +
-        content.substring(minIndex-10);
-        
-    return trimmedText.replace(new RegExp('\\b(?:' + terms.join('|') + ')\\b', 'ig'),
-                               (w) => `<mark>${escape(w)}</mark>`);
-}
-
-router.get('/search', iv.validateGET({ q: 'string' }, { json: true }), (req, res) => {
-    const results = searchIndex.index.search(req.query.q);
-    const data = [];
-    for (let i = 0; i < Math.min(5, results.length); i++) {
-        const result = results[i];
-        data.push({
-            url: result.ref,
-            score: result.score,
-            highlight: highlightSearch(result.ref, result.matchData.metadata)
-        });
-    }
-    
-    res.cacheFor(86400);
-    res.json({
-        result: 'ok',
-        data
-    });
+router.get('/:page', (req, res, next) => {
+    if (req.params.page.endsWith('.md'))
+        res.redirect(301, '/doc/' + req.params.page);
+    else
+        next();
 });
-
-router.use('/thingpedia-api', express.static(path.join(__dirname, '../doc/thingpedia-api')));
 
 module.exports = router;
