+++ views/doc_base.pug
@@ -35,7 +35,7 @@ block content
     div.modal-dialog
       div.modal-content
         div.modal-header.text-center
-          h4.modal-title= _("Thingpedia Documentation Search")
+          h4.modal-title= _("Documentation Search")
         div.modal-body
           form(action='#', autocomplete='off')
             div.input-group#docsearch-container
