+++ views/stanford/about_get_involved.pug
@@ -37,7 +37,7 @@ block content
                 Do you own a device, or know of a service that you want to control through Almond?
                 You can add it to Thingpedia, and have Almond immediately learn about it!
               p.
-                Check out our #[a(href='/thingpedia/developers/thingpedia-intro.md') developer documentation],
+                Check out our #[a(href='/doc/thingpedia-intro.md') developer documentation],
                 and when you're ready, #[a(href='/user/request-developer') apply to our developer program].
 
             div.col-md-6
