+++ views/stanford/about_index.pug
@@ -206,7 +206,7 @@ block content
               p
                 = _("Almond uses a state-of-the-art natural language understanding model. Almond's deep learning model allows it to understand more complex commands across more domains than any other assistant on the market: just train Almond with pairs of sentences and programs, and Almond will learn.")
               p
-                != _("We are continuously improving Almond, and building tools to allow others to extend Almond. We have developed a tool called <a href='/thingpedia/developers/genie-intro.md'>Genie</a> that allows to generate training data in new domains from scratch. Using Genie, we have found that Almond understands user's input with 68% accuracy - a marked improvement over the previous best known result. Furthermore, with little effort we were able to build a specialized skill for <a href='https://almond.stanford.edu/thingpedia/devices/by-id/com.spotify'>Spotify</a>, add natural language to access control, and extend Almond with aggregates (sum, count, average, etc.).")
+                != _("We are continuously improving Almond, and building tools to allow others to extend Almond. We have developed a tool called <a href='/doc/genie-intro.md'>Genie</a> that allows to generate training data in new domains from scratch. Using Genie, we have found that Almond understands user's input with 68% accuracy - a marked improvement over the previous best known result. Furthermore, with little effort we were able to build a specialized skill for <a href='https://almond.stanford.edu/thingpedia/devices/by-id/com.spotify'>Spotify</a>, add natural language to access control, and extend Almond with aggregates (sum, count, average, etc.).")
               p
                 != _("As academics, our research is open-source, and all technology is freely available to the public. Anyone can leverage our algorithm in their product or in their own research. <a href='https://oval.cs.stanford.edu'>Learn more about our research</a> and <a href='/about/use-almond'>how you can use Almond's technology</a>.")
