+++ views/stanford/about_use_almond.pug
@@ -40,7 +40,7 @@ block content
             div.col-md-6.hanging-icon-block#almond-tech-open-api
               h4= _("API")
               p.
-                Almond provides an #[a(href='/thingpedia/developers/my-api.md') open API] that you can
+                Almond provides an #[a(href='/doc/my-api.md') open API] that you can
                 use from your service or device directly. This allows you to integrate with Almond's natural
                 language technology without deploying Almond on your own server.
 
@@ -50,7 +50,7 @@ block content
               p.
                 Almond's natural language model, called #[i LUInet], is freely available to anybody to use.
                 Additionally, you can train your own natural language understanding model using Almond's technology,
-                dataset and resource. Check out our #[a(href='/thingpedia/developers/genie-intro.md') documentation]
+                dataset and resource. Check out our #[a(href='/doc/genie-intro.md') documentation]
                 to learn more.
